normed

+++ srcs/stack_init.c
@@ -12,7 +12,7 @@
 
 #include "ft_push_swap.h"
 
-void				print_all(t_list *a, t_list *b)
+void	print_all(t_list *a, t_list *b)
 {
 	printf("list a :");
 	print_list(a);
@@ -22,23 +22,23 @@ void				print_all(t_list *a, t_list *b)
 	printf("\n");
 }
 
-void				print_list(t_list *a)
+void	print_list(t_list *a)
 {
 	t_list	*tmp;
-	
+
 	tmp = a;
 	if (tmp)
 	{
 		while (tmp && tmp->next)
 		{
-			printf("%i ",tmp->content);
+			printf("%i ", tmp->content);
 			tmp = tmp->next;
 		}
-		printf("%i\n",(int)tmp->content);
+		printf("%i\n", tmp->content);
 	}
 }
 
-void		addstack(int size, char **table, t_list **list)
+void	addstack(int size, char **table, t_list **list)
 {
 	t_list	*tmp;
 	int		content;
